Use double instead of float for generated types
Qt uses double mostly. In QML all reals are stored in double precision.
DBus also works with doubles.
Clients also use double more often than float. So it would be better if
generator will work with double instead of float to avoid conversion
from double to float and prevent precision lost.

+++ src/model/FaceliftConversion.h
@@ -417,18 +417,18 @@ struct TypeHandler<int> : public TypeHandlerBase
 
 
 template<>
-struct TypeHandler<float> : public TypeHandlerBase
+struct TypeHandler<double> : public TypeHandlerBase
 {
-    typedef float QMLType;
+    typedef double QMLType;
 
-    static QString toString(const float &v)
+    static QString toString(const double &v)
     {
         return QString::number(v);
     }
 
-    static float fromVariant(const QVariant &variant)
+    static double fromVariant(const QVariant &variant)
     {
-        return variant.toFloat();
+        return variant.toDouble();
     }
 };
